+++ DSCResources/IisServer/IisServer.schema.psm1
@@ -108,7 +108,11 @@ Configuration IisServer
     $stigDataObject    = [STIG]::New( $StigVersion, $OrgSettings, $technology,
                                           $technologyRole, $technologyVersion, $Exception,
                                           $SkipRuleType, $SkipRule )
+<<<<<<< HEAD
                                           
+=======
+
+>>>>>>> origin/2.4.0.0
     #### BEGIN DO NOT MODIFY ####
     # $StigData is used in the resources that are dot sourced below
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseDeclaredVarsMoreThanAssignments",'')]