+++ DSCResources/IisSite/IisSite.schema.psm1
@@ -118,6 +118,10 @@ Configuration IisSite
     $stigDataObject    = [STIG]::New( $StigVersion, $OrgSettings, $technology,
                                           $technologyRole, $technologyVersion, $Exception,
                                           $SkipRuleType, $SkipRule )
+<<<<<<< HEAD
+=======
+
+>>>>>>> origin/2.4.0.0
     #### BEGIN DO NOT MODIFY ####
     # $StigData is used in the resources that are dot sourced below
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseDeclaredVarsMoreThanAssignments",'')]