+++ Tests/Integration/DSCResources/IisServer.integration.tests.ps1
@@ -15,7 +15,11 @@ try
     foreach ($stig in $stigList)
     {
         [xml] $dscXml = Get-Content -Path $stig.Path
+<<<<<<< HEAD
         
+=======
+
+>>>>>>> origin/2.4.0.0
         Describe "IIS Server $($stig.StigVersion) mof output" {
 
             It 'Should compile the MOF without throwing' {
@@ -54,7 +58,7 @@ try
             Context 'WebConfigurationPropertyRule' {
                 $hasAllSettings = $true
                 $dscXml = $dscXml.DISASTIG.WebConfigurationPropertyRule.Rule
-                $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[xWebConfigProperty\]"}
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[xWebConfigProperty\]"}
 
                 foreach ($setting in $dscXml)
                 {
@@ -74,7 +78,7 @@ try
             Context 'MimeTypeRule' {
                 $hasAllSettings = $true
                 $dscXml = $dscXml.DISASTIG.MimeTypeRule.Rule
-                $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[xIisMimeTypeMapping\]"}
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[xIisMimeTypeMapping\]"}
 
                 foreach ($setting in $dscXml)
                 {
@@ -92,10 +96,17 @@ try
         }
 
         Describe "IIS Server $($stig.StigVersion) Single SkipRule/RuleType mof output" {
+<<<<<<< HEAD
             
             $SkipRule = Get-Random -InputObject $dscXml.DISASTIG.MimeTypeRule.Rule.id
             $SkipRuleType = "IisLoggingRule"
             
+=======
+
+            $SkipRule     = Get-Random -InputObject $dscXml.DISASTIG.MimeTypeRule.Rule.id
+            $SkipRuleType = "IisLoggingRule"
+
+>>>>>>> origin/2.4.0.0
             It 'Should compile the MOF without throwing' {
                 {
                     & "$($script:DSCCompositeResourceName)_config" `
@@ -107,29 +118,52 @@ try
                         -SkipRuleType $SkipRuleType `
                 } | Should not throw
             }
+<<<<<<< HEAD
            
+=======
+
+>>>>>>> origin/2.4.0.0
             #region Gets the mof content
             $configurationDocumentPath = "$TestDrive\localhost.mof"
             $instances = [Microsoft.PowerShell.DesiredStateConfiguration.Internal.DscClassCache]::ImportInstances($configurationDocumentPath, 4)
             #endregion
+<<<<<<< HEAD
            
             Context 'Skip check' {
+=======
+
+            Context 'Skip check' {
+
+>>>>>>> origin/2.4.0.0
                 #region counts how many Skips there are and how many there should be.
                 $dscIisLoggingRuleXml = $dscXml.DISASTIG.IisLoggingRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "pass"}
                 $expectedSkipRuleCount = ($($dscIisLoggingRuleXml.Count) + $($SkipRule.Count))
                 $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"}
                 #endregion
+<<<<<<< HEAD
+=======
+
+>>>>>>> origin/2.4.0.0
                 It "Should have $expectedSkipRuleCount Skipped settings" {
                     $dscMof.count | Should Be $expectedSkipRuleCount
                 }
             }
         }
+<<<<<<< HEAD
         
         Describe "IIS Server $($stig.StigVersion) Multiple SkipRule/RuleType mof output" {
             
             $SkipRule = Get-Random -InputObject $dscXml.DISASTIG.WebConfigurationPropertyRule.Rule.id -Count 2
             $SkipRuleType = @('MimeTypeRule','IisLoggingRule')
            
+=======
+
+        Describe "IIS Server $($stig.StigVersion) Multiple SkipRule/RuleType mof output" {
+
+            $SkipRule     = Get-Random -InputObject $dscXml.DISASTIG.WebConfigurationPropertyRule.Rule.id -Count 2
+            $SkipRuleType = @('MimeTypeRule','IisLoggingRule')
+
+>>>>>>> origin/2.4.0.0
             It 'Should compile the MOF without throwing' {
                 {
                     & "$($script:DSCCompositeResourceName)_config" `
@@ -141,11 +175,16 @@ try
                         -SkipRuleType $SkipRuleType `
                 } | Should not throw
             }
+<<<<<<< HEAD
             
+=======
+
+>>>>>>> origin/2.4.0.0
             #region Gets the mof content
             $configurationDocumentPath = "$TestDrive\localhost.mof"
             $instances = [Microsoft.PowerShell.DesiredStateConfiguration.Internal.DscClassCache]::ImportInstances($configurationDocumentPath, 4)
             #endregion
+<<<<<<< HEAD
             
             Context 'Skip check' {
                
@@ -156,11 +195,26 @@ try
                 $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"}
                 #endregion
                 
+=======
+
+            Context 'Skip check' {
+
+                #region counts how many Skips there are and how many there should be.
+                $dscMimeTypeRuleXml = $dscXml.DISASTIG.MimeTypeRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "Pass"}
+                $dscIisLoggingRuleXml = $dscXml.DISASTIG.IisLoggingRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "Pass"}
+
+                $expectedSkipRuleCount = ($($dscMimeTypeRuleXml.Count) + $($dscIisLoggingRuleXml.count) + $($SkipRule.Count))
+
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"}
+                #endregion
+
+>>>>>>> origin/2.4.0.0
                 It "Should have $expectedSkipRuleCount Skipped settings" {
                     $dscMof.count | Should Be $expectedSkipRuleCount
                 }
             }
         }
+<<<<<<< HEAD
 
         Describe "IIS Server $($stig.StigVersion) Exception mof output"{
             
@@ -182,6 +236,8 @@ try
                 } | Should not throw
             }
         }
+=======
+>>>>>>> origin/2.4.0.0
     }
 }
 #endregion Tests